mac: fix first real compile errors from an actual Xcode toolchain
- ScreenCapture.swift: SCContentFilter.pointPixelScale is Float, not
  CGFloat -- couldn't multiply directly against contentRect.width/height
  (CGFloat). Explicit CGFloat(...) conversion.
- AccessibilityReader.swift: kAXLinkRole isn't actually exposed by the
  ApplicationServices Swift overlay (unlike the other kAX*Role globals
  used alongside it) -- switched the whole clickableRoles set to literal
  role strings ("AXButton" etc.) to remove the dependency entirely rather
  than fix one constant and risk another missing one on the next round.
  Also: this Swift version treats CFTypeRef -> AXUIElement/AXValue
  downcasts as statically infallible and errors on the conditional/forced
  forms (as?/as!) as redundant -- switched to plain `as` in
  focusedEditableElement() and boundsAttr().
- ActionDispatcher.swift / OcrEngine.swift: cleaned up two warnings
  (unused MainActor.run result, a no-op conditional downcast now that
  VNRecognizeTextRequest.results is already correctly typed) while in
  the area.

First real build feedback since this was written blind from a sandbox
with no macOS toolchain -- confirms the "Known risk areas" framing in
mac/README.md was warranted. Grepped for other as?/as! AX*/CF* patterns
across the rest of Sources/ to preempt a repeat of this class of error;
none found.

+++ mac/Sources/GameAutopilotMac/Capture/ScreenCapture.swift
@@ -89,7 +89,7 @@ actor ScreenCapture {
         // Default output size does NOT auto-match the window/display --
         // must set explicitly (accounting for Retina backing scale) or
         // frames come back downscaled, hurting OCR/mark accuracy.
-        let scale = filter.pointPixelScale
+        let scale = CGFloat(filter.pointPixelScale)
         config.width = max(1, Int((filter.contentRect.width * scale).rounded()))
         config.height = max(1, Int((filter.contentRect.height * scale).rounded()))
